+++ ExperimentalEnemyInteractions/EnemyPatches/SandSpiderAIPatch.cs
@@ -88,8 +88,7 @@ namespace NaturalSelection.EnemyPatches
             }
             if (__instance.IsOwner)
             {
-                List<EnemyAI> temoList = NaturalSelectionLib.NaturalSelectionLib.globalEnemyLists[type];
-                spiderData.enemiesInLOSDictionary = new Dictionary<EnemyAI, float>(LibraryCalls.GetEnemiesInLOS(__instance, ref temoList, 80f, 15, 2f));
+                spiderData.enemiesInLOSDictionary = new Dictionary<EnemyAI, float>(LibraryCalls.GetEnemiesInLOS(__instance, 80f, 15, 2f));
             }
 
             if (spiderData.enemiesInLOSDictionary.Count > 0)