+++ ExperimentalEnemyInteractions/Generics/LibraryCalls.cs
@@ -1,7 +1,6 @@
-using System;
 using System.Collections.Generic;
 using BepInEx.Logging;
-using NaturalSelection.EnemyPatches;
+using UnityEngine;
 
 namespace NaturalSelection.Generics;
 
@@ -67,10 +66,16 @@ public class LibraryCalls
         NaturalSelectionLib.NaturalSelectionLib.FilterEnemySizes(ref importEnemySizeDict, enemySizes, instance, inverseToggle);
     }
 
-    static public Dictionary<EnemyAI, float> GetEnemiesInLOS(EnemyAI instance, ref List<EnemyAI> importEnemyList, float width = 45f, int importRange = 0, float proximityAwareness = -1)
+    static public Dictionary<EnemyAI, float> GetEnemiesInLOS(EnemyAI instance, ref List<EnemyAI> importEnemyList, float width = 45f, int importRange = 0, float proximityAwareness = -1, Vector3? importEyePosition = null)
     {
         if (debugLibraryCalls && debugSpam && debugTriggerFlag) Script.Logger.Log(LogLevel.Info,"Called library GetEnemiesInLOS!");
-        return NaturalSelectionLib.NaturalSelectionLib.GetEnemiesInLOS(instance, ref importEnemyList, width, importRange, proximityAwareness);
+        return NaturalSelectionLib.NaturalSelectionLib.GetEnemiesInLOS(instance, ref importEnemyList, width, importRange, proximityAwareness, importEyePosition);
     }
-    
+
+    static public Dictionary<EnemyAI, float> GetEnemiesInLOS(EnemyAI instance, float width = 45f, int importRange = 0, float proximityAwareness = -1, Vector3? importEyePosition = null)
+    {
+        if (debugLibraryCalls && debugSpam && debugTriggerFlag) Script.Logger.Log(LogLevel.Info, "Called library GetEnemiesInLOS!");
+        return NaturalSelectionLib.NaturalSelectionLib.GetEnemiesInLOS(instance, width, importRange, proximityAwareness, importEyePosition);
+    }
+
 }